# docs/api-ledger/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "BuyerAnalyticsResource" : {
        "properties" : {
          "companyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "completedPayments" : {
            "format" : "int64",
            "type" : "integer"
          },
          "monthlySpending" : {
            "items" : {
              "$ref" : "#/components/schemas/MonthlyAmount"
            },
            "type" : "array"
          },
          "pendingPayments" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalOrders" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalSpent" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "BuyerCompanyResource" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "contactEmail" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "ruc" : {
            "type" : "string"
          },
          "sector" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CreateBuyerCompanyResource" : {
        "properties" : {
          "address" : {
            "minLength" : 1,
            "type" : "string"
          },
          "contactEmail" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "phone" : {
            "minLength" : 1,
            "type" : "string"
          },
          "ruc" : {
            "minLength" : 1,
            "pattern" : "\\d{11}",
            "type" : "string"
          },
          "sector" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "address", "contactEmail", "name", "phone", "ruc", "sector" ],
        "type" : "object"
      },
...
      "FuelRequestResource" : {
        "properties" : {
          "buyerCompanyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "deliveryDate" : {
            "format" : "date",
            "type" : "string"
          },
          "equipmentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "fuelProductId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "fuelType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "productName" : {
            "type" : "string"
          },
          "providerId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "quantity" : {
            "format" : "double",
            "type" : "number"
          },
          "rejectionReason" : {
            "type" : "string"
          },
          "source" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "APPROVED", "REJECTED" ],
            "type" : "string"
          },
          "unit" : {
            "type" : "string"
          },
          "unitPrice" : {
            "format" : "double",
            "type" : "number"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "MonthlyAmount" : {
        "properties" : {
          "amount" : {
            "format" : "double",
            "type" : "number"
          },
          "month" : {
            "type" : "string"
          },
          "monthIndex" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "NotificationResource" : {
        "properties" : {
          "createdAt" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          },
          "read" : {
            "type" : "boolean"
          },
          "referenceId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "enum" : [ "NEW_REQUEST", "REQUEST_PENDING", "ORDER_ACCEPTED", "ORDER_REJECTED", "ORDER_DISPATCHED", "ORDER_DELIVERED", "ORDER_CONFIRMED", "ORDER_CANCELLED", "DELIVERY_DISPATCHED", "DELIVERY_COMPLETED", "DELIVERY_FAILED", "PAYMENT_RECEIVED", "PAYMENT_COMPLETED", "PAYMENT_REFUNDED", "GENERAL" ],
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PromotedUser" : {
        "properties" : {
          "roles" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token JWT Bearer para autenticar solicitudes a la API",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "support@primefuel.com",
      "name" : "Soporte de FullTank",
      "url" : "https://primefuel.com/support"
    },
    "description" : "API de FullTank para gestionar abastecimiento, flota, entregas, inventario y pagos por organizaciÃ³n.",
    "license" : {
      "name" : "Apache 2.0 (licencia)",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "FullTank Platform",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/analytics/buyers/{companyId}" : {
      "get" : {
        "description" : "Devuelve métricas agregadas de la empresa indicada cuando el usuario pertenece a ella.",
        "operationId" : "getBuyerAnalytics",
        "parameters" : [ {
          "in" : "path",
          "name" : "companyId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerAnalyticsResource"
                }
              }
            },
            "description" : "Indicadores de la empresa compradora devueltos."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerAnalyticsResource"
                }
              }
            },
            "description" : "El usuario no pertenece a la empresa compradora solicitada."
          }
        },
        "summary" : "Consultar indicadores de empresa compradora",
        "tags" : [ "Analítica" ]
      }
    },
...
    "/api/v1/buyer-companies" : {
      "get" : {
        "description" : "Devuelve todos los perfiles registrados. Solo está disponible para administradores de plataforma.",
        "operationId" : "getAllBuyerCompanies",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BuyerCompanyResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se devuelve la lista de empresas compradoras."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BuyerCompanyResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "El usuario no cuenta con autoridad ROLE_ADMIN."
          }
        },
        "summary" : "Listar empresas compradoras",
        "tags" : [ "Empresas compradoras" ]
      },
      "post" : {
        "description" : "Persiste un nuevo perfil de empresa compradora. La ruta está disponible sin autenticación y no crea usuarios ni membresías.",
        "operationId" : "createBuyerCompany",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateBuyerCompanyResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Perfil de empresa compradora creado."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Se produjo un error inesperado al guardar el perfil."
          }
        },
        "summary" : "Registrar empresa compradora",
        "tags" : [ "Empresas compradoras" ]
      }
    },
    "/api/v1/buyer-companies/{companyId}" : {
      "get" : {
        "description" : "Devuelve el perfil indicado cuando pertenece al tenant comprador autenticado.",
        "operationId" : "getBuyerCompanyById",
        "parameters" : [ {
          "in" : "path",
          "name" : "companyId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerCompanyResource"
                }
              }
            },
            "description" : "Perfil de empresa compradora devuelto."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerCompanyResource"
                }
              }
            },
            "description" : "La empresa compradora no pertenece al usuario autenticado."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerCompanyResource"
                }
              }
            },
            "description" : "La empresa compradora no existe."
          }
        },
        "summary" : "Consultar empresa compradora por identificador",
        "tags" : [ "Empresas compradoras" ]
      },
      "put" : {
        "description" : "Reemplaza los campos editables del perfil identificado. Requiere que el usuario pertenezca a esa empresa.",
        "operationId" : "updateBuyerCompany",
        "parameters" : [ {
          "in" : "path",
          "name" : "companyId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateBuyerCompanyResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerCompanyResource"
                }
              }
            },
            "description" : "Perfil de empresa compradora actualizado."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerCompanyResource"
                }
              }
            },
            "description" : "La empresa compradora no pertenece al usuario autenticado."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerCompanyResource"
                }
              }
            },
            "description" : "La empresa compradora no existe."
          }
        },
        "summary" : "Actualizar empresa compradora",
        "tags" : [ "Empresas compradoras" ]
      }
    },
...
    "/api/v1/fuel-requests/{requestId}" : {
      "get" : {
        "description" : "Devuelve la solicitud indicada cuando el usuario pertenece a la empresa compradora o al distribuidor correspondiente.",
        "operationId" : "findById",
        "parameters" : [ {
          "in" : "path",
          "name" : "requestId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "Solicitud de combustible devuelta."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FuelRequestResource"
                }
              }
            },
            "description" : "La solicitud no existe o no pertenece al usuario autenticado."
          }
        },
        "summary" : "Consultar solicitud de combustible por identificador",
        "tags" : [ "Solicitudes de combustible heredadas" ]
      }
    },
...
    "/api/v1/notifications/buyer/{companyId}" : {
      "get" : {
        "description" : "Devuelve las notificaciones del usuario asociado a la empresa compradora cuando el tenant autenticado es propietario.",
        "operationId" : "getNotificationsByBuyer",
        "parameters" : [ {
          "in" : "path",
          "name" : "companyId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/NotificationResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se devuelve la lista, que puede estar vacía si la empresa no tiene usuario asociado."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/NotificationResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "La empresa solicitada no pertenece al tenant autenticado."
          }
        },
        "summary" : "Listar notificaciones por empresa compradora",
        "tags" : [ "Notificaciones" ]
      }
    },
...
    "/api/v2/admin/users/{userId}/promote" : {
      "post" : {
        "description" : "Añade ROLE_ADMIN a la cuenta sin retirar los roles existentes. Solo puede hacerlo un administrador de plataforma.",
        "operationId" : "promote",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotedUser"
                }
              }
            },
            "description" : "La cuenta tiene el rol ROLE_ADMIN."
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotedUser"
                }
              }
            },
            "description" : "El usuario no tiene autoridad de administrador de plataforma."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromotedUser"
                }
              }
            },
            "description" : "La cuenta no existe."
          }
        },
        "summary" : "Otorgar rol de administrador de plataforma",
        "tags" : [ "Administración de plataforma" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Servidor de la API",
    "url" : "http://localhost:8080"
  } ],
...
}